we're banished from a time in a fallen land
(Forgot password).....

To a life beyond the stars

In your darkest dreams see to believe
Our destiny is time
And endlessly we'll all be free tonight!

And on the wings of a dream
So far beyond reality
All alone in desperation
Now the time has gone

Lost inside you'll never find
Lost within my own mind
Day after day this misery must go on!

So far away we wait for the day
For the lives all so wasted and gone
We feel the pain of a lifetime lost in a thousand days
Through the fire and the flames we carry on!

[guitar solo]

Wo-o-oah,wo-o-oah,who-o-o-o-oah!
Wo-o-oah,wo-o-oa,who-o-o-o-oah!

[guitar solo]

Now here we stand with their blood on our hands
We fought so hard, now can we understand
I'll break the seal of this curse if I possibly can
For freedom of every man!

So far away we wait for the day
For the lives all so wasted and gone
We feel the pain of a lifetime lost in a thousand days
Through the fire and the flames we carry on!